refactor: move cart component to it's own folder
