+++ src/App.tsx
@@ -6,7 +6,7 @@ import ProductDetails from './pages/ProductDetails';
 import Products from './pages/Products/Products';
 import Header from './components/Header/Header';
 import Footer from './components/Footer/Footer';
-import Cart from './pages/Cart';
+import Cart from './pages/Cart/Cart';
 import loadPageTheme from './utils/loadPageTheme';
 import 'bootstrap/dist/css/bootstrap.min.css';
